Feat/pluggable-autocleaning-step-1 (#221)
Makes management of lowcode JLisp interpreter, and autocleaning methods swappable like styling and postprocessing.

Decided to sideline the full implementation for now because it is too involved... and polars makes some autocleaning functionality very difficult, particularly comparing original to modfified across different dtypes. This makes it much more difficult to color and add tooltips to the resulting dataframe based on modifications.

Nothing is integrated into the main flow yet. This just makes the autocleaning module which will be integrated